YD-662 Encrypt creation time
The creation time is now moved to the UserPrivate. On the User entity, we still have a date, but that's rounded to monthly dates, to reduce identification risk (correlating the creation date/time with something else on the anonymized side of things). The rounding does not occur for users that use the app for less than 30 days, to allow building the dashboard with statistics. The data of a short-term user is anyway less interesting.

Next step: create an upgrade step for the existing users.