YD-664 Send Firebase notification for system msg (#574)
* YD-664 Send Firebase notification for system msg

The Spring Batch code uses direct JPA instead of using the Spring services. Because of that, the Firebase message needs to be sent explicitly. Added that now.

* YD-664 Added integration test

Apparently, I didn't run the integration tests yet, as the system message test failed before I extended it with Firebase assertions: the message had to be saved before it could be given to the Firebase service. This is fixed now.

To enable asserting this, I moved the test controller from the app service to Core, so all four services have it now. Similarly, I moved the related methods of AppService.groovy to Service.groovy.