YD-668 Do not allow leading zeros in mobile nr
Till now, we allowed numbers like +3106..., with a zero (could even be more than one) between the country code and the rest of the number. It looks like the mobile number confirmation logic generally doesn't work in that case, as only one user has a confirmed number of this format. However, if they use it to fetch a new device request, that fails. Similar issues exist when inviting buddies.

With this change, such numbers are not allowed anymore.