YD-653 Do not show activities from before joining Yona (#573)
* YD-653 No activities prior to user/buddy creation

Paging back through the user or buddy activities will not show inactivity entries for dates before the user was created or the buddy relationship was established.

In this stages, all tests pass again. Except for ActivityTest."Delete goal after update", no tests have been updated functionally. Several tests cut off the activities on mandatory goals. This is not realistic, as users will always have these, so I want to update the tests to stop doing that. I already had to do that for the "Delete goal after update" test, as that would otherwise have been very superficial.

Along with these changes, the following was done:
* Introduce a "is test server" setting. Operations created for integration tests are now dependent on this setting. To prevent from accidentally enabling this in production, it is made incompatible with enabling SMS.

Another commit(s) will follow with:
* Integration tests updates (stop truncating the mandatory goals)
* A test for a user with activities from multiple buddies with different "last status change times"
* Any changes that surface in a final review of the change set

* YD-653 Made the tests more real-world

The tests were cutting off the mandatory goals, but in real life that never happens. The tests now don't do that anymore.

Updating the tests revealed an issue: The class WeekActivityDto was creating day inacitivty instances for all days of the week, also those that preceded the account creation date. This is fixed now.

It would be good to extend the tests with a global assertion that verifies whether the week activity overview indeed matches the expected values. Now the week activity overview could contain unexpected activities.

* YD-653 Added test to exercise buddy behavior

This revealed a bug where 'with buddies' returned a day activity that shouldn't be returned (dating from before the buddy relationship was established)

* YD-653 Added tests for boundary conditions

Fixed several missing boundary checks

* YD-653 Assert test server in clearLastFirebaseMessage

* YD-653 Added testServer to Helm and scripts

Along with this:
* Made HibernateStatsTest to work again. It was unknowingly broken for some time.
* Added a test in BuddyService.updateLastStatusChangeTime to assert the last status change time does not predate the user creation
* Extended the PropertiesValidator with logic to ensure the flag enableHibernateStatsAllowed can only be enabled on a test server

* YD-653 Code improvements flagged by SonarQube

* Removed unused logger from PropertiesValidator.java
* Removed unused parameter userCreationDate from ActivityService.getTotalPageableItems
* Restructured complex for loop in WeekActivityDto.createRequiredInactivityDays into a filtered stream

* YD-653 Processed review comment

Added explanatory comment to updateBuddy

* YD-653 Resolved review comments

* In ActivityService.getUserDayActivityOverviewWithBuddies, assert the user has at least one buddy. Added assertions to the related integration test.
* Swapped the parameter order of ActivityService.assertDateNotTooEarly, in line with the order in JUnit
* Moved ActivityService.limitInterval into Interval.limit
* ActivityService.getInterval now asserts the endDate against the earliestPossibleDate, giving a more logical error message.
* Removed UserAnonymizedDto.getOldestGoalCreationTime and UserAnonymizedDto.getOldestVersionOfGoal as these are now unused.
* Added unit tests to the max operations of TimeUtil

Along with this:
* Assert a regular user update does not include a creation date change

* YD-653 Removed the isChanged methods

from the DTOs. The behavior was very specific, too specific for a DTO. It is now reimplemented in the services that need to do the comparison.