YD-667 Don't use variable for next URL
The samplers for fetching the first and next page are being executed in random order, so the value for the 'next' URL might not yet be available when fetching the next page, as the next page might be fetched before the first one (random order!).

Now the next page simply adds "page=1" to the path.