YD-667 Set user creation time
For goals, it was already possible to set the creation time when creating the goal. This is now possible for users too, provided the server is configured as test server (this check was missing for goals).

The load test makes use of this new feature, to enable paging through day and week activities.