YD-669 Pass app info all the way (#580)
* YD-669 Pass app info all the way

It now goes from the app service to invoked services (e.g. the analysis service) and also to the async thread that sends the Firebase message, so it can be included in the error logging there.

Along with this:
* Improved the error logging for errors returned from an invoked service

* YD-669 Consistent arg order params, headers

We used to have methods that take just an argument "parameters" and overwriting methods that use "headers, parameters", both optional. This lead to inconsistencies. Sometimes the single given argument is "parameters", sometimes it's "headers". Now the parameter order is always "parameters, headers".

* YD-669 Include the token in the error message

When an Firebase exception occurs.

* YD-669 Addressed SonarQube review comment

Replaced lambda call with method reference

* YD-671 Corrected path

'infrastructure' only prefixes the path for the regression test server, not for the other servers.

* YD-669 Addressed review comments:
* Better name for HeadersHolder
* Better variable names in FirebaseTest
* Added some comments

* YD-669 Explained pass through headers approach

In class-level comments.

* YD-669 Addressed review comments

* Spelling correction in comment in HeadersServerInterceptor
* Corrected private method PassThroughHeadersHolder.storeIfPresent (it didn't use the name parameter in all necessary places)