YD-677 Resolved multithreading issue pass-through headers
The issue was that we were trying to inject a thread-scoped bean into a singleton bean. However, the injection on a singleton bean happens only once, so the same PassThroughHeaders instance was used on all threads.

Added a test that reproduces the issue, created a straightforward fix and removed the thread-scope stuff.