YD-665 Use cached user anonymized for device info
The anonymized user entity is cached. It includes the anonymized device info, which is often needed. This is now reused, so the get....First calls are occasionally more inefficient, but the get...Second calls are a lot more efficient.

This leads to the following new numbers: the test run loaded 1094 (vs 1248) entities through 252 (vs 259) statements and 86 (vs 97) queries