YD-687 Address concurrency issues
Concurrent requests to resend the PIN reset confirmation code caused a 500 error. This is addressed now. The issue was caused by stale data: some entities were already loaded into the session before acquiring the update lock on the user entity. If these entities were updated in a previous transaction that the load-with-lock request is waiting for, then these updates are not loaded into the session, causing the current session to continue with stale data. This is also what occurred with YD-620. To prevent this, the session is now cleared prior to acquiring the update lock.

A significant part of the change set is caused by using the user entity rather than the DTO. This allows reusing the locked user entity, it simplifies the code a bit and makes it a bit faster.

Along with this:
* PinResetRequestTest now asserts the presence/absence of the resendPinResetConfirmationCodeUrl
* UserTest now asserts proper behavior for trying to delete and get nonexisting users