YD-687 Correct goal update scenarios
The previous change broke scenarios where the goal with activities was updated. The issue is that originally, we initialized the GoalIdMapping using the DTOs but then we started using the entities. The DTOs have a collection goals which actually contains goalsIncludingHistoryItems. Because of that, the tests that use activities on updated goals now fail, as the history items are not available in the mapping. This is corrected now.

* The goal ID mapping is now initialized including the history items
* Several more goal collections that actually contain "goals including history items" are now renamed
* The user entity is locked prior to making goal changes
* The behavior for calculating the set of goals including history items (which was already based on entities) is now moved to the UserAnonymized entity
* The entity loading statistics were recalculated: the situation worsened considerably. Before this PR is merged, I want to see whether that can be recovered.