YD-683 Don't fail pin reset when user is removed (#593)
* YD-683 Don't fail pin reset when user is removed

If a user decides to remove or overwrite their account while waiting for a pin reset confirmation code, the batch job generating the reset code should not fail. This is now addressed.

Along with this:
* Removed unused attribute messageSupplierCaptor from BuddyDeviceChangeMessageDtoTest

* YD-634 Handle PIN reset within lock on user

This prevents concurrency issues.

* YD-687 Address concurrency issues

Concurrent requests to resend the PIN reset confirmation code caused a 500 error. This is addressed now. The issue was caused by stale data: some entities were already loaded into the session before acquiring the update lock on the user entity. If these entities were updated in a previous transaction that the load-with-lock request is waiting for, then these updates are not loaded into the session, causing the current session to continue with stale data. This is also what occurred with YD-620. To prevent this, the session is now cleared prior to acquiring the update lock.

A significant part of the change set is caused by using the user entity rather than the DTO. This allows reusing the locked user entity, it simplifies the code a bit and makes it a bit faster.

Along with this:
* PinResetRequestTest now asserts the presence/absence of the resendPinResetConfirmationCodeUrl
* UserTest now asserts proper behavior for trying to delete and get nonexisting users

* YD-688 Prevent forced resend of PIN reset confirmation code

Till now, it was possible to force sending a confirmation code though the delay window was not expired. This is addressed now.

* YD-683 Processed review comment

The change in an earlier commit was accidentally removed. This is restored now.

* YD-683 Corrected name goalsIncludingHistoryItems

The class UserPrivateDataBaseDto has a collection goals that actually contains the goals including history items. That was never made explicit in the name and now that lead to a failure. Renamed it now.

* YD-687 Correct goal update scenarios

The previous change broke scenarios where the goal with activities was updated. The issue is that originally, we initialized the GoalIdMapping using the DTOs but then we started using the entities. The DTOs have a collection goals which actually contains goalsIncludingHistoryItems. Because of that, the tests that use activities on updated goals now fail, as the history items are not available in the mapping. This is corrected now.

* The goal ID mapping is now initialized including the history items
* Several more goal collections that actually contain "goals including history items" are now renamed
* The user entity is locked prior to making goal changes
* The behavior for calculating the set of goals including history items (which was already based on entities) is now moved to the UserAnonymized entity
* The entity loading statistics were recalculated: the situation worsened considerably. Before this PR is merged, I want to see whether that can be recovered.

* YD-687 Optimized entity loading

Regained pretty much all of the previous loss by using the cached user anonymized entity. Along with this corrected a slight mistake in UserAddService (accidentally used a static method on an instance).

* YD-687 Removed unused import

Flagged by SonarQube