YD-686 Added length check for activity comments (#600)
* YD-686 Added length check for activity comments

Earlier, there was no explicit check. Testing proved I could make a comment with 237 *s. In reality, this limit might vary due to the encryption. To stay on the safe side and still give plenty of room for a comment, the maximum is now set to 200 characters.

Along with this:
* Corrected a few error messages by taking out double curly braces
* Removed two duplicated error messages

* YD-686 Processed review comments

* In messages.properties, replace back ticks with ticks
* Extracted addComment as a method in ActivityCommentTest