YD-695 Switch to Read Committed (#612)
* YD-695 Debug logging added

Temporarily, will be removed when implementing the fix

* YD-695 Switch to Read Committed

Now use Read Committed as transaction isolation level.

Introduced OverwriteUserTest.'Concurrent requests cause no errors'. This consistently failed before switching the isolation level and now consistently passes. Generalized the method postNTimes of PinResetRequestTest.

* YD-695 Now set COMMITTED rather than UNCOMMITTED

The comment on the ticket had a copy/paste error that was repeated in the code. Corrected it now. The test still consistently passes.

* YD-695 Refactored concurrent POST code

Along with this ensured the AsyncHTTPBuilder is shut down after use.

* YD-695 Refactored async handler

Made it more generic, for use in future tests